add annotation for cases that use oneauthtestapp
